Move filter.js to views/
